PROBLEM: INT64_MIN was still being returned from one branch of persitence code SOLUTION: Change the to the expected INT32_MIN

+++ src/db/measurements/measurement.cc
@@ -146,7 +146,7 @@ select_measurements_averages (
                 result.getRow (0).getValue (0).get (topic_id);
             }
             else if (result.size () == 0) {
-                last_timestamp = INT64_MIN;
+                last_timestamp = INT32_MIN;
                 ret.rv = 2;
                 return ret;
             }
